Move legacy mapped TGI index to separate file
